Get distance function (#120)
* added getDistanceFrom function to vecsim index

* increasing float_vector_normalize working range

* added tests

* test fix(?)

* test fix 2(?)

* test fix 3(?)

* test fix 4(?)

* bug fix in IP optimization function,
that caused rounding results

* test fix 5(!)

* another space test update

+++ src/VecSim/utils/vec_utils.cpp
@@ -47,9 +47,9 @@ int cmpVecSimQueryResultByScore(const VecSimQueryResult *res1, const VecSimQuery
 }
 
 void float_vector_normalize(float *x, size_t dim) {
-    float sum = 0;
+    double sum = 0;
     for (size_t i = 0; i < dim; i++) {
-        sum += x[i] * x[i];
+        sum += (double)x[i] * (double)x[i];
     }
     float norm = sqrt(sum);
     if (norm == 0)
